refactor(ado-script): promote PR helpers to shared/ for upcoming contributors
Stage 0 of the execution-context contributor build-out (plan.md, session
artifact). Pure file move; no behavioural change. Moves four helper
modules from scripts/ado-script/src/exec-context-pr/ into
scripts/ado-script/src/shared/ so the upcoming pipeline, ci-push,
schedule, and workitem contributors can reuse them without
fragmenting the workspace with an exec-context-common/ sibling.

Files moved (with git mv, preserving history):
- git.ts        (bearerEnv + runGit / gitOk wrappers)
- merge-base.ts (synthetic-merge detection + progressive-deepening fetch)
- validate.ts   (identifier allowlist regexes + sanitizeForPrompt)
- prompt.ts     (success / failure prompt-fragment writers + appender)

And their __tests__/ siblings.

shared/index.ts gains four new barrels (git, mergeBase, validate,
prompt). exec-context-pr/index.ts now imports from ../shared/
instead of ./. Doc references in docs/ado-script.md and
src/compile/extensions/exec_context/pr.rs updated to match.

Validation:
- cargo build / cargo test (2013 tests, 0 failures) / cargo clippy clean
- npm run codegen / typecheck / test (283 tests) / build / test:smoke (6 tests)
