+++ scripts/ado-script/src/shared/__tests__/git.test.ts
