+++ scripts/ado-script/src/shared/__tests__/prompt.test.ts
