+++ scripts/ado-script/src/shared/__tests__/validate.test.ts
