+++ scripts/ado-script/src/shared/git.ts
