+++ scripts/ado-script/src/shared/index.ts
@@ -3,3 +3,11 @@ export * as vso from "./vso-logger.js";
 export * as envFacts from "./env-facts.js";
 export * as policy from "./policy.js";
 export * as adoClient from "./ado-client.js";
+// Promoted from exec-context-pr/ during Stage 0 of the contributor
+// build-out so upcoming contributors (`pipeline`, `ci-push`,
+// `workitem`, ...) can reuse them without fragmenting the workspace
+// with an `exec-context-common/` sibling. See plan.md "Stage 0".
+export * as git from "./git.js";
+export * as mergeBase from "./merge-base.js";
+export * as validate from "./validate.js";
+export * as prompt from "./prompt.js";
