+++ scripts/ado-script/src/shared/merge-base.ts
