+++ scripts/ado-script/src/shared/prompt.ts
