+++ scripts/ado-script/src/shared/validate.ts
